backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz>=2023.3
pytest>=7.3.1
httpx>=0.24.0